QDir: fix int overflow
This caused reverse order of session items in qt creator. Introduced
in ba287c55ef179b073482453298f513992b54c11e.

+++ src/corelib/io/qdir.cpp
@@ -219,7 +219,7 @@ bool QDirSortItemComparator::operator()(const QDirSortItem &n1, const QDirSortIt
     if ((qt_cmp_si_sort_flags & QDir::DirsLast) && (f1->item.isDir() != f2->item.isDir()))
         return !f1->item.isDir();
 
-    int r = 0;
+    qint64 r = 0;
     int sortBy = (qt_cmp_si_sort_flags & QDir::SortByMask)
                  | (qt_cmp_si_sort_flags & QDir::Type);
